fixed type

+++ src/zs_error.cc
@@ -80,8 +80,9 @@ void throw_builtin_identifier_check_failed(Lisp_ptr p){
   throw_zs_error(p, "arg is not identifier!");
 }
 
-void throw_builtin_range_check_failed(int max, int passed){
-  throw_zs_error({}, "index is out-of-bound ([0, %d), supplied %d\n",
+void throw_builtin_range_check_failed(size_t max, int passed){
+  // The 'z' specifier is a C99 feature, included in C++11.
+  throw_zs_error({}, "index is out-of-bound ([0, %zu), supplied %d\n",
                  max, passed);
 }
